+++ packages/components/fields/date-range-field/src/date-range-field.story.js
@@ -9,9 +9,9 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
-import * as icons from '../../icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
+import * as icons from '../../../icons';
 import DateRangeField from './date-range-field';
 
 storiesOf('Components|Fields', module)
