+++ packages/components/fields/date-range-field/src/date-range-field.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DateRangeField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = ['2018-09-20', '2018-09-24'];
