+++ packages/components/fields/date-range-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './date-range-field';
+
+export { default as version } from './version';
