--- src/components/fields/date-range-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './date-range-field';
